requirements/prod.txt
Django==4.2.1
django-cleanup==8.1.0
python-dotenv==1.0.1
sorl-thumbnail==12.10.0
pillow==10.2.0
django-ckeditor==6.7.1